Add Menu component
A new Menu component was added. This was created to encapsulate the logic for navigation menus in the core module. This component includes a simple HTML file, TypeScript file and its respective test file. A CSS file was also included for styling options, however, it is still empty.

In addition to the Menu component, the package-lock.json was altered to handle dependencies. It now includes specific version information, ensuring that installed versions of packages are kept consistent across multiple environments.